Add a test case for WFLY-12982 "MP FT: CDI contexts not available in @Timeout methods" and restructure tests
