+++ testsuite/integration/microprofile/src/test/java/org/wildfly/test/integration/microprofile/faulttolerance/context/asynchronous/AsyncService.java
@@ -19,11 +19,10 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.test.integration.microprofile.faulttolerance.async.requestcontext;
+package org.wildfly.test.integration.microprofile.faulttolerance.context.asynchronous;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Future;
-
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
